+++ Examples/TensorDerivedImage.cxx
@@ -42,7 +42,7 @@ int main( int argc, char *argv[] )
   typedef itk::ImageFileWriter<ScalarImageType> WriterType;
 
   //typedef itk::Vector<PixelType, 6>             TensorType;
-  typedef itk::SymmetricSecondRankTensor< double, 3 >  TensorType; 
+  typedef itk::SymmetricSecondRankTensor< float, 3 >  TensorType; 
   typedef itk::Image<TensorType, 3>             TensorImageType;
   typedef itk::ImageFileReader<TensorImageType> ReaderType;
   typedef itk::RGBPixel< float >                ColorPixelType;
